+++ release/plugins/main/tag/index.es6.js
@@ -1,1 +1,1 @@
-export { tag } from './tag.es6.js';
+export { TAG_GROUP_NAME, tag } from './tag.es6.js';