+++ release/plugins/main/tag/index.js
@@ -6,4 +6,5 @@ var tag = require('./tag.js');
 
 
 
+exports.TAG_GROUP_NAME = tag.TAG_GROUP_NAME;
 exports.tag = tag.tag;