+++ release/plugins/main/tag/tag.d.ts
@@ -1,4 +1,9 @@
 import { type HandlerPlugin } from 'innet';
+export interface TagGroup {
+    name: string;
+    tags: string[];
+}
+export declare const TAG_GROUP_NAME = "x-tagGroups";
 export interface TagProps {
     children?: any;
     /**
@@ -6,7 +11,9 @@ export interface TagProps {
      * [CommonMark syntax](https://spec.commonmark.org) MAY be used for rich text representation.
      * */
     description?: string;
-    /** The name of the tag. */
+    /** A name of the tag group. */
+    group?: string;
+    /** A name of the tag. */
     name: string;
 }
 export declare const tag: HandlerPlugin;