+++ src/hooks/useSearch/useSearch.ts
@@ -5,7 +5,7 @@ export function useSearch <D> (): D {
   const action = useAction()
 
   if (!action) {
-    useThrow('<{type}> MUST be in <request> or <fallback>')
+    useThrow('<{type}> MUST be in <return> or <preset>')
   }
 
   return action.search as D