+++ src/plugins/request/header/header.ts
@@ -12,7 +12,7 @@ export const header: HandlerPlugin = () => {
   const res = useResponse()
 
   if (!res) {
-    useThrow('<{type}> MUST be in <request> or <fallback>')
+    useThrow('<{type}> MUST be in <return> or <preset>')
   }
 
   const {