+++ src/plugins/schema/integer/integer.ts
@@ -79,10 +79,10 @@ export const integer: HandlerPlugin = () => {
   } = useProps<IntegerProps>() || {}
   const schema = useSchemaType('integer', {
     ...props,
-    const: props.const !== undefined ? Number(props.const) : undefined,
     default: defaultValue !== undefined ? Number(defaultValue) : undefined,
     example: example !== undefined ? Number(example) : undefined,
     examples: examples?.map(Number),
+    value: props.value !== undefined ? Number(props.value) : undefined,
     values: values && getArrayValues(values, Number),
   })