+++ release/handler/handler.d.ts
@@ -1,7 +1,6 @@
+import { type Plugin } from 'innet';
 import { type JSXElement } from '@innet/jsx';
-import { arraySync, async } from '@innet/utils';
 import { type AnyProps, type ApiProps, type ArrayProps, type BinaryProps, blacklist, type BlacklistProps, type BodyProps, type BooleanProps, cms, type CmsProps, type ContactProps, type CookieProps, type DateProps, type DtsProps, type EndpointProps, type EnvProps, type ErrorProps, type FieldProps, file, type FileProps, type HeaderProps, type HostProps, type IntegerProps, type LicenseProps, type NullProps, type NumberProps, type ObjectProps, type ParamProps, preset, type PresetProps, protection, type ProtectionProps, type ProxyProps, type RedirectProps, type ResponseProps, type ReturnProps, type ServerProps, type StringProps, type SuccessProps, type TagProps, type TupleProps, type UiProps, type UuidProps, type VariableProps, whitelist, type WhitelistProps } from '../plugins';
-export declare const arrayPlugins: (typeof arraySync)[];
 export declare const JSXPlugins: {
     any: import("innet").HandlerPlugin;
     api: import("innet").HandlerPlugin;
@@ -44,9 +43,7 @@ export declare const JSXPlugins: {
     variable: import("innet").HandlerPlugin;
     whitelist: typeof whitelist;
 };
-export declare const fnPlugins: import("innet").Plugin[];
-export declare const objectPlugins: import("innet").Plugin[];
-export declare const promisePlugins: (typeof async)[];
+export declare const objectPlugins: Plugin[];
 export declare const handler: import("innet").Handler;
 declare global {
     namespace JSX {