+++ release/handler/index.js
@@ -7,8 +7,5 @@ var handler = require('./handler.js');
 
 
 exports.JSXPlugins = handler.JSXPlugins;
-exports.arrayPlugins = handler.arrayPlugins;
-exports.fnPlugins = handler.fnPlugins;
 exports.handler = handler.handler;
 exports.objectPlugins = handler.objectPlugins;
-exports.promisePlugins = handler.promisePlugins;